+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - 🛠️ Fixed nav bar not highlighting active page.
 - ⚙️ Configured CSS-loader for modules in production mode, this is why the :global() was not working in production.
+- ⚙️ Replaced custom popup events with creating new instances of the popup element.
 
 ### **2.0.0 (8 May, 2025)**